+++ kafka-pub/src/app/repositories.ts
@@ -1,3 +1,4 @@
+import { TaskRepository } from "@/repositories";
 import type { Provider } from "@nestjs/common";
 
-export const Repositories: Provider[] = [];
+export const Repositories: Provider[] = [TaskRepository];